Split app from index
